+++ src/ts/index.ts
@@ -161,6 +161,7 @@ import CodeHighlightTabs from "./components/extensions/codehighlight/CodeHighlig
 import InlineCodeHighlight from "./components/extensions/codehighlight/InlineCodeHighlight";
 import Notification from "./components/extensions/notifications/Notification";
 import NotificationProvider from "./components/extensions/notifications/NotificationProvider";
+import NotificationContainer, {appNotifications} from "./components/extensions/notifications/NotificationContainer";
 import NavigationProgress from "./components/extensions/nprogress/NavigationProgress";
 import NavigationProgressProvider from "./components/extensions/nprogress/NavigationProgressProvider";
 import MantineProvider from "./components/styles/MantineProvider";
@@ -268,6 +269,7 @@ export {
     NavigationProgressProvider,
     Notification,
     NotificationProvider,
+    NotificationContainer,
     NumberFormatter,
     NumberInput,
     Overlay,
@@ -338,4 +340,5 @@ export {
     YearPickerInput,
     RichTextEditor,
     TypographyStylesProvider,
+    appNotifications
 };